Reorganize LensAlg alias stuff
