+++ src/main/scala/LensAlgHom.scala
@@ -1,6 +1,8 @@
 package org.hablapps.statelesser
 
 import scalaz._
+import shapeless._, ops.hlist._
+import naturally._
 
 trait LensAlgHom[Alg[_[_], _], P[_], A] {
   type Q[_]
@@ -28,5 +30,26 @@ object LensAlgHom {
       val alg = alg2
       def apply() = app2
     }
+
+  implicit def genLensAlg[
+        Rev <: HList, Ctx <: HList : Reverse.Aux[Rev, ?], S, A](implicit 
+      ev: DeepLens.Aux[S, Ctx, A]): GetEvidence[Rev, LensAlg[State[S, ?], A]] =
+    GetEvidence(LensAlg(ev()))
+
+  implicit def genLensAlgHom[
+        Rev <: HList, Ctx <: HList : Reverse.Aux[Rev, ?], 
+        Alg[_[_], _], S, A](implicit 
+      ge: GetEvidence[HNil, Alg[State[A, ?], A]],
+      dl: DeepLens.Aux[S, Ctx, A])
+      : GetEvidence[Rev, LensAlgHom[Alg, State[S, ?], A]] =
+    GetEvidence(LensAlgHom(ge(), dl()))
+
+  trait Syntax {
+    implicit class LensSyntax[P[_], A](la: LensAlg[P, A]) {
+      def get: P[A] = la.fold(_.get)
+      def set(a: A): P[Unit] = la.fold(_.put(a))
+      def modify(f: A => A): P[Unit] = la.fold(_.modify(f))
+    }
+  }
 }
